feat[214]: added sorting of users by name and date
