+++ public/locales/en/nav_bar.json
@@ -1,10 +1,10 @@
 {
-    "home": "Home",
-    "create": "Create",
-    "my_profile": "My Profile",
-    "messenger": "Messenger",
-    "search": "Search",
-    "statistics": "Statistics",
-    "favorites": "Favorites",
-    "log_out": "Log Out"
+  "home": "Home",
+  "create": "Create",
+  "my_profile": "My Profile",
+  "messenger": "Messenger",
+  "search": "Search",
+  "statistics": "Statistics",
+  "favorites": "Favorites",
+  "log_out": "Log Out"
 }
